feat: add timestamps and status info, disable IPv6, remove unneeded code
There was previously no verbosity built-in, so now timestamps and status
types will be prepended to all messages. The status types are:
- INIT: initializing
- GRAB: downloading new content
- NEWR: upgrading quality of existing content
- EXIT: exiting
- WARN: warn user of something

IPv6 does not seem to work for network requests to scryfall.com, so
it will now be forcefully disabled upon program execution.

Many variables were refactored but remain the same.

Some calls and functions were removed due to being unnecessary. This
will not change the user experience.